Add option to skip simplification
Summary
-------
This diff adds a parameter to the constructor of the `Fraction` class
called `$shouldSimplify`. When set to false it skips the last step of
calling the `simplify()` function.

The parameter is defaulted to `true` so there is no change in existing
behavior.

Test Plan
--------
See new unit test.

Questions
---------
- My other thought was making `simplify()` protected so that we could
  subclass it and remove the simplification functionality in a subclass.

Context
-------
We want to be able to validate that a string is "fraction-like", but
keep the format that the string was originally in.