Strictly type Fraction class
In the process of doing this I discovered one bug where `toFloat` would
not always return a float if the numerator and denominator, when
divided, came out to a whole number. Now it will return the float
version of that whole number.

I also deleted a couple test cases that I think are irrelevant in a
scalar type hint world.

Note that this would make this library compatible with PHP 7 only. If
that's not acceptable for the new version I'm fine leaving this change
out.